[Rust SDK] Add SDK boilerplate with utils
