+++ rust/Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "moving-average",
     "post-processor",
     "processor",
+    "sdk",
     "server-framework",
 ]
 
@@ -22,6 +23,7 @@ rust-version = "1.70"
 processor = { path = "processor" }
 server-framework = { path = "server-framework" }
 aptos-moving-average = { path = "moving-average" }
+aptos-processor-sdk = { path = "sdk" }
 
 anyhow = "1.0.62"
 async-trait = "0.1.53"
@@ -42,6 +44,7 @@ diesel = { version = "2.1", features = [
 diesel-async = { version = "0.4", features = ["postgres", "bb8", "tokio"] }
 diesel_migrations = { version = "2.1.0", features = ["postgres"] }
 diesel_async_migrations = { git = "https://github.com/niroco/diesel_async_migrations", rev = "11f331b73c5cfcc894380074f748d8fda710ac12" }
+duration-str = "0.7.0"
 enum_dispatch = "0.3.12"
 field_count = "0.1.1"
 futures = "0.3.24" # Previously futures v0.3.23 caused some consensus network_tests to fail. We now pin the dependency to v0.3.24.
@@ -73,6 +76,7 @@ sha2 = "0.9.3"
 sha3 = "0.9.1"
 strum = { version = "0.24.1", features = ["derive"] }
 tempfile = "3.3.0"
+thiserror = "1"
 toml = "0.7.4"
 tracing-subscriber = { version = "0.3.17", features = ["json", "env-filter"] }
 tokio = { version = "1.21.0", features = ["full"] }
