+++ rust/processor/Cargo.toml
@@ -13,8 +13,13 @@ repository = "https://github.com/aptos-labs/aptos-core"
 rust-version = { workspace = true }
 
 [dependencies]
-anyhow = { workspace = true }
+# Internal dependencies.
 aptos-moving-average = { workspace = true }
+aptos-processor-sdk = { workspace = true }
+server-framework = { workspace = true }
+
+# External dependencies.
+anyhow = { workspace = true }
 aptos-protos = { workspace = true }
 async-trait = { workspace = true }
 base64 = { workspace = true }
@@ -41,7 +46,6 @@ prost-types = { workspace = true }
 regex = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-server-framework = { workspace = true }
 sha2 = { workspace = true }
 sha3 = { workspace = true }
 strum = { workspace = true }
