+++ rust/processor/src/models/default_models/block_metadata_transactions.rs
@@ -6,10 +6,8 @@
 #![allow(clippy::unused_unit)]
 
 use super::transactions::Transaction;
-use crate::{
-    schema::block_metadata_transactions,
-    utils::util::{parse_timestamp, standardize_address},
-};
+use crate::{schema::block_metadata_transactions, utils::util::standardize_address};
+use aptos_processor_sdk::utils::parse_timestamp;
 use aptos_protos::{
     transaction::v1::BlockMetadataTransaction as BlockMetadataTransactionPB,
     util::timestamp::Timestamp,
