+++ rust/processor/src/processors/token_v2_processor.rs
@@ -32,10 +32,11 @@ use crate::{
             clean_data_for_db, execute_with_better_error, get_chunks, MyDbConnection, PgDbPool,
             PgPoolConnection,
         },
-        util::{get_entry_function_from_user_request, parse_timestamp, standardize_address},
+        util::{get_entry_function_from_user_request, standardize_address},
     },
 };
 use anyhow::bail;
+use aptos_processor_sdk::utils::parse_timestamp;
 use aptos_protos::transaction::v1::{transaction::TxnData, write_set_change::Change, Transaction};
 use async_trait::async_trait;
 use diesel::{pg::upsert::excluded, result::Error, ExpressionMethods};
