+++ rust/sdk/src/lib.rs
@@ -1,0 +1,7 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod utils;
+
+// Re-export other crates devs will need to write processors.
+pub use aptos_protos;
